+++ libs/@blockprotocol/core/src/core-embedder-handler.ts
@@ -1,5 +1,5 @@
-import { CoreHandler } from "./core-handler";
-import { EmbedderInitMessage, Message, MessageContents } from "./types";
+import { CoreHandler } from "./core-handler.js";
+import type { EmbedderInitMessage, Message, MessageContents } from "./types.js";
 
 /**
  * Implements the Block Protocol Core Specification for embedding applications.