+++ libs/block-template-custom-element/CHANGELOG.md
@@ -1,5 +1,14 @@
 # block-template-custom-element
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies [[`345ceaf3`](https://github.com/blockprotocol/blockprotocol/commit/345ceaf39abfdfac11e87cc7355746def3b888fb)]:
+  - @blockprotocol/graph@0.3.4
+  - block-scripts@0.3.5
+  - mock-block-dock@0.1.10
+
 ## 0.4.5
 
 ### Patch Changes